feat:기본값초기화
