--- Study_Branch/Champion.cs
@@ -1,25 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace Study_Branch
-{
-    internal class Champion
-    {
-        protected int hp = 100;
-        protected int mp = 30;
-        private int attackpoint;
-        private int ammor;
-
-        public Champion()
-        {
-            this.hp = hp;
-            this.mp = mp;
-            this.attackpoint = attackpoint;
-            this.ammor = ammor;
-
-        }
-    }
-}
