+++ Study_Branch/LOL/Program.cs
@@ -1,0 +1,10 @@
+﻿namespace Study_Branch.LOL
+{
+    internal class Program
+    {
+        static void Main(string[] args)
+        {
+            Console.WriteLine("Hello, World!");
+        }
+    }
+}
