+++ Study_Branch/LOL/Unit.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace Study_Branch
+namespace Study_Branch.LOL
 {
     internal class Unit
     {
