--- Study_Branch/Program.cs
@@ -1,10 +1,0 @@
-﻿namespace Study_Branch
-{
-	internal class Program
-	{
-		static void Main(string[] args)
-		{
-			Console.WriteLine("Hello, World!");
-		}
-	}
-}
